fix(ui): center element handling across all backends + VM arithmetic fixes
- Remove parser center→col transformation, preserve center as independent widget
- Iced: nested container approach for centering with max-width constraint
- Vue: remove w-full override, add mx-auto for horizontal centering with max-w-*
- Rust code gen: special handling for View::center(child) builder pattern
- Fix VM arithmetic opcodes (ADD/SUB/MUL/DIV/NEG) nanbox 2-slot f64 handling
- Fix SET_GENERIC_FIELD pop_nv→pop_arith_operand for nanbox safety
- Fix stale input_values after on_with_input in dynamic UI
- Add style utility helpers (w_full, h_full, iced_length)

+++ crates/auto-lang/src/ui/vm_bridge.rs
@@ -809,7 +809,7 @@ impl VmBridge {
 fn eval_aura_expr_to_value(expr: &AuraExpr) -> Value {
     match expr {
         AuraExpr::Int(i) => Value::Int(*i as i32),
-        AuraExpr::Float(f) => Value::Float(*f as f64),
+        AuraExpr::Float(f) => Value::Double(*f as f64),
         AuraExpr::Bool(b) => Value::Bool(*b),
         AuraExpr::Literal(s) => Value::str(s),
         AuraExpr::StateRef(name) => {
